Add files via upload
Bakery management system is a system that is especially build for providing the details and bills of bakery items like cakes, pastry and cookies on computer. This system helps the manager as well as customer to manage their ordered product easily. It consumes less time than writing a bill by hand and it makes it easy for the customer and manager to view the items with their rates and other basic information on computer.
We used classes, objects, inheritance, polymorphism and files for update and delete record. Class diagram of this system is also mentioned.                                                                           In this project, first of all we make the class named “Bakery” which stores the information of customer. It has a pure virtual function of “void cash” which is used for printing of bill. Due to the virtual function in this class, the class becomes an abstract class. By using the concept of inheritance, three derived classes are derived from the base class “Bakery ". The three classes are “cakes, cookies, pastry”.
In cake class, we have its additional functions i.e. “void show_cakes” which is used for displaying all the varieties of cakes and “void get_name” which is used for entering the user’s choice in order to get the item of that choice. The function of base class “Bakery” that stores the person’s information are again called by “this”, (pointer name), pointer in order to inherit the “get_name” function of base class" Bakery in function of “get_name” of the cake class. The virtual function of base class, void cash, is over-ride in this derived class to print the bill of cakes for the customer.
In cookies class, we have its additional functions i.e. “void show_cookies”; which is used for displaying all the varieties of cookies and “void get_name” which is used for entering the user’s choice in order to get the item of that choice, the function of base class that stores the person’s information are again called by “this” pointer in order to inherit the “get_name” function of base class in function of “get_name” of the cookies class. The virtual function of base class, void cash, is over-ride in this derived class to print the bill of cookies for the customer.
In pastry class, we have its additional functions i.e. “void show_pastry” which is used for displaying all the varieties of pastry and “void get_name” which is used for entering the user’s choice in order to get the item of that choice, the function of base class that stores the person’s information are again called by “this” pointer in order to inherit the “get_name” function of base class in function of “get_name” of the pastry class. The virtual function of base class, void cash, is over-ride in this derived class to print the bill of pastry for the customer.
In main function (), we make the objects of inherited classes that are cakes, cookies and pastry. And as we can’t make an object of base class because it is an abstract class. So, according to the concept of polymorphism, we declare three pointers of base class Bakery that are “*b1, *b2, *b3” to point the derived classes. By applying loop, inside the loop we put condition of if, else if and else. Under these conditions we specified the choice of user for the specific function and according to that choice the respected functions under the condition is called. “Void cash” function is called by the pointer of base class e.g. “b1->cash ()” for the functionality of respected class and remaining functions of inherited classes are called by the objects of that class using dot operator e.g. “P. show_pastry () “. At the end, all the functions are called and display all the required things.